changed navigation on pages and added features to admin to add products in db
