+++ actions.ts
@@ -1,6 +1,7 @@
 "use server";
 
-import { createUser, deleteUser } from "./api";
+import { createUser, deleteUser, BASE_URL } from "./api";
+import { revalidatePath } from "next/cache";
 
 export async function createUserAction(formData: FormData) {
   const { name, email } = Object.fromEntries(formData);
@@ -10,4 +11,24 @@ export async function createUserAction(formData: FormData) {
 
 export async function deleteUserAction(id: number) {
   await deleteUser(id);
+}
+
+export async function addProduct(formData: any) {
+  try {
+    const response = await fetch(BASE_URL + "/api/add-product", {
+      method: "POST",
+      cache: "no-store",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(formData),
+    });
+    revalidatePath("/");
+    if (response.ok) {
+      return await response.json();
+    } else {
+      throw new Error(`Error: ${response.statusText}`);
+    }
+  } catch (error) {
+    console.error("Error submitting form:", error);
+    throw new Error("Submission failed");
+  }
 }
