+++ messages/ka.json
@@ -1,7 +1,7 @@
 {
   "Navigation": {
     "home": "მთავარი",
-    "About": "ჩვენს შესახებ",
+    "Products": "პროდუქტები",
     "Blogs": "ბლოგები",
     "Contact": "კონტაქტი",
     "Profile": "პროფილი",
