+++ src/app/[locale]/(dashboard)/admin/add/page.tsx
@@ -1,0 +1,11 @@
+import AddProductForm from "../../../../components/AddProductForm";
+
+const AddProduct = () => {
+  return (
+    <div>
+      <AddProductForm />
+    </div>
+  );
+};
+
+export default AddProduct;
